# openapi.yaml
openapi: 3.0.0
info:
  title: SPK
  version: 1.0.0
  description: ini untuk spk
servers:
  - url: '{{BASE_URL}}'
  - url: '{{PROCESSOR_URL}}'
paths:
...
  /calculate:
    post:
      summary: calculate
      responses:
        '200':
          description: '200'
          headers:
            date:
              schema:
                type: string
              example: Thu, 14 May 2026 19:21:11 GMT
            server:
              schema:
                type: string
              example: uvicorn
            content-length:
              schema:
                type: integer
              example: 43969
        '422':
          description: 422 (file not uploaded)
          headers:
            date:
              schema:
                type: string
              example: Thu, 14 May 2026 19:52:28 GMT
            server:
              schema:
                type: string
              example: uvicorn
            content-length:
              schema:
                type: integer
              example: 89
      tags:
        - Processor
      description: |-
        Untuk request antar service, gak usah terlalu dipedulikan.

        body ini contoh saja
  /upload:
    post:
      summary: upload
      parameters:
        - name: write
          in: query
          schema:
            type: boolean
          example: true
      responses:
        '200':
          description: '200'
          headers:
            Content-Security-Policy:
              schema:
                type: string
              example: >-
                default-src 'self';base-uri 'self';font-src 'self' https:
                data:;form-action 'self';frame-ancestors 'self';img-src 'self'
                data:;object-src 'none';script-src 'self';script-src-attr
                'none';style-src 'self' https:
                'unsafe-inline';upgrade-insecure-requests
            Cross-Origin-Opener-Policy:
              schema:
                type: string
              example: same-origin
            Origin-Agent-Cluster:
              schema:
                type: string
              example: '?1'
            Referrer-Policy:
              schema:
                type: string
              example: no-referrer
            Strict-Transport-Security:
              schema:
                type: string
              example: max-age=31536000; includeSubDomains
            X-Content-Type-Options:
              schema:
                type: string
              example: nosniff
            X-DNS-Prefetch-Control:
              schema:
                type: string
              example: 'off'
            X-Download-Options:
              schema:
                type: string
              example: noopen
            X-Frame-Options:
              schema:
                type: string
              example: SAMEORIGIN
            X-Permitted-Cross-Domain-Policies:
              schema:
                type: string
              example: none
            X-XSS-Protection:
              schema:
                type: integer
              example: 0
            Vary:
              schema:
                type: string
              example: Origin
            Access-Control-Allow-Credentials:
              schema:
                type: boolean
              example: true
            Content-Length:
              schema:
                type: integer
              example: 36
            ETag:
              schema:
                type: string
              example: W/"24-0Kwf0r3R04ob+DiSKhz+NUykBgM"
            Date:
              schema:
                type: string
              example: Thu, 14 May 2026 20:01:54 GMT
            Connection:
              schema:
                type: string
              example: keep-alive
            Keep-Alive:
              schema:
                type: string
              example: timeout=5
        '500':
          description: '500'
          headers:
            Content-Security-Policy:
              schema:
                type: string
              example: >-
                default-src 'self';base-uri 'self';font-src 'self' https:
                data:;form-action 'self';frame-ancestors 'self';img-src 'self'
                data:;object-src 'none';script-src 'self';script-src-attr
                'none';style-src 'self' https:
                'unsafe-inline';upgrade-insecure-requests
            Cross-Origin-Opener-Policy:
              schema:
                type: string
              example: same-origin
            Origin-Agent-Cluster:
              schema:
                type: string
              example: '?1'
            Referrer-Policy:
              schema:
                type: string
              example: no-referrer
            Strict-Transport-Security:
              schema:
                type: string
              example: max-age=31536000; includeSubDomains
            X-Content-Type-Options:
              schema:
                type: string
              example: nosniff
            X-DNS-Prefetch-Control:
              schema:
                type: string
              example: 'off'
            X-Download-Options:
              schema:
                type: string
              example: noopen
            X-Frame-Options:
              schema:
                type: string
              example: SAMEORIGIN
            X-Permitted-Cross-Domain-Policies:
              schema:
                type: string
              example: none
            X-XSS-Protection:
              schema:
                type: integer
              example: 0
            Vary:
              schema:
                type: string
              example: Origin
            Access-Control-Allow-Credentials:
              schema:
                type: boolean
              example: true
            Set-Cookie:
              schema:
                type: string
              example: >-
                refresh_token=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOjEsImlhdCI6MTc3ODc4OTE0MiwiZXhwIjoxNzc5MzkzOTQyfQ.2z8BkufrjGY4p1R394hUK9t0wO0gpj6bF2LRqhC2v8k;
                Max-Age=604800; Path=/; Expires=Thu, 21 May 2026 20:05:42 GMT;
                HttpOnly; SameSite=Lax
            Content-Length:
              schema:
                type: integer
              example: 88
            ETag:
              schema:
                type: string
              example: W/"58-nIKI0aqqTYwrXOiNyswh5Qeb/us"
            Date:
              schema:
                type: string
              example: Thu, 14 May 2026 20:06:05 GMT
            Connection:
              schema:
                type: string
              example: keep-alive
            Keep-Alive:
              schema:
                type: string
              example: timeout=5
      tags:
        - 'TOPSIS data '
      description: param nya kalo gak pake write=true berarti dia gak nulis ke database
...
  /data:
    get:
      summary: get data
      parameters:
        - name: page
          in: query
          schema:
            type: integer
          example: 1
        - name: pageSize
          in: query
          schema:
            type: integer
          example: 10
      responses:
        '200':
          description: '200'
          headers:
            Content-Security-Policy:
              schema:
                type: string
              example: >-
                default-src 'self';base-uri 'self';font-src 'self' https:
                data:;form-action 'self';frame-ancestors 'self';img-src 'self'
                data:;object-src 'none';script-src 'self';script-src-attr
                'none';style-src 'self' https:
                'unsafe-inline';upgrade-insecure-requests
            Cross-Origin-Opener-Policy:
              schema:
                type: string
              example: same-origin
            Origin-Agent-Cluster:
              schema:
                type: string
              example: '?1'
            Referrer-Policy:
              schema:
                type: string
              example: no-referrer
            Strict-Transport-Security:
              schema:
                type: string
              example: max-age=31536000; includeSubDomains
            X-Content-Type-Options:
              schema:
                type: string
              example: nosniff
            X-DNS-Prefetch-Control:
              schema:
                type: string
              example: 'off'
            X-Download-Options:
              schema:
                type: string
              example: noopen
            X-Frame-Options:
              schema:
                type: string
              example: SAMEORIGIN
            X-Permitted-Cross-Domain-Policies:
              schema:
                type: string
              example: none
            X-XSS-Protection:
              schema:
                type: integer
              example: 0
            Vary:
              schema:
                type: string
              example: Origin
            Access-Control-Allow-Credentials:
              schema:
                type: boolean
              example: true
            Content-Length:
              schema:
                type: integer
              example: 14021
            ETag:
              schema:
                type: string
              example: W/"36c5-/d4wcJO6qN9NICDwScIH3r0F4cA"
            Date:
              schema:
                type: string
              example: Thu, 14 May 2026 20:19:33 GMT
            Connection:
              schema:
                type: string
              example: keep-alive
            Keep-Alive:
              schema:
                type: string
              example: timeout=5
      tags:
        - 'TOPSIS data '